refactor(ui): extract shared time utilities to lib/time.ts
- Add formatDuration, formatTime, formatShortDate to lib/time.ts
- Update SessionTimeline to use shared utilities
- Reduces code duplication across dashboard components
